requirements.txt
# Crop Delineation Pipeline — Python Dependencies

# Core
torch>=2.0.0
torchvision>=0.15.0
rasterio>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
Pillow>=9.0.0
labelme

# Segmentation
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0

# Geospatial
shapely>=2.0.0
pyproj>=3.5.0
pystac-client>=0.7.0
planetary-computer>=1.0.0
